Move frontend model into models
